refactor: 컴포넌트 폴더 생성 후 style 분리
