+++ src/components/CardList/CardList.style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+export const Grid = styled.div`
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 12px;
+  width: 100%;
+  justify-items: center;
+`
