+++ src/components/ShowMoreButton/ShowMoreButton.style.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled'
+
+export const Button = styled.button`
+  width: 100%;
+  margin: 32px 0 0 0;
+  padding: 16px 0;
+  border: none;
+  border-radius: 12px;
+  cursor: pointer;
+  transition: background 0.2s;
+`
