--- src/index.css
@@ -1,1 +1,0 @@
-/* src/assets/index.css */
