+++ src/main.tsx
@@ -3,7 +3,6 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { ThemeProvider } from '@emotion/react'
 import theme from '@/styles/theme'
-import '@/index.css'
 import App from '@/App'
 
 createRoot(document.getElementById('root')!).render(
