+++ src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import NavigationBar from "@/components/NavigationBar"
+import NavigationBar from "@/components/NavigationBar/NavigationBar"
 import SelectFriendSection from "@/sections/SelectFriendSection"
 import CategorySection from "@/sections/CategorySection"
 import RankingSection from "@/sections/RankingSection"
